+++ dev_build/lib/src/mixin/package_io.dart
@@ -4,4 +4,4 @@
 library;
 
 export 'package:dev_build/src/package/recursive_pub_path.dart'
-    show recursiveActions, recursivePubPath;
+    show recursiveActions, recursivePubPath, iteratePubPath;